+++ src/routes/Router.tsx
@@ -9,18 +9,19 @@ import Calendar from '../pages/Calendar';
 import MyPage from '../pages/MyPage';
 import Layout from '../components/Layout';
 import RoomDetail from '../pages/RoomDetail';
-import Materials from '../pages/Materials';
-import Homework from '../pages/Homework';
 import CounselingDiary from '../pages/CounselingDiary';
 import Payment from '../pages/Payment';
 import Statistics from '../pages/Statistics';
 import CreateRoom from '../pages/CreateRoom';
 import KakaoOauth from '../components/KakaoOauth';
-import CreateHomework from '../pages/CreateHomework';
-import HomeworkDetail from '../pages/HomeworkDetail';
 import EditRoom from '../pages/EditRoom';
 import SignupForm from '../pages/SignupForm';
 import SignupComplete from '../pages/SignupComplete';
+import Materials from '../pages/Material/Materials';
+import CreateMaterials from '../pages/Material/CreateMaterials';
+import Homework from '../pages/Homework/Homework';
+import CreateHomework from '../pages/Homework/CreateHomework';
+import HomeworkDetail from '../pages/Homework/HomeworkDetail';
 
 export const router = createBrowserRouter([
   {
